test(architecture): give the shared justfile parser its own home
#1444 extended the justfile parser foundation_slice_test.go already
owned rather than writing a second one, which was the right call and
left the parser serving two guardrails from inside the file for one of
them. doc.go described that file as only the test-foundation rule, so
the second caller was invisible to anyone reading the package map.

The recipe graph — justfileName, justRecipes and justRecipe with their
methods, justHeaderLine, parseJustfile and parseJustfileText — moves to
justfile_parse_test.go, with a header saying what both callers want out
of it and why a second parser would be two definitions of what a recipe
is rather than duplicated lines. This is the convention the package
already has for a shared reader, after native_constants_test.go and
repo_walk_test.go, and doc.go gains the matching entry.

No behaviour changes and no assertion moves. The fixture-driven test
#1444 added stays in justfile_doc_test.go, because what it asserts on is
recipeIsListed and summaryFault — the summary rule, not the parser.
withoutComments and readRepoFile stay in foundation_slice_test.go too:
both serve the CI-workflow reader beside them and neither is specific to
reading a justfile.
